print_requirements.txt
Package           Version
----------------- -------
lxml              5.4.0
MouseInfo         0.1.3
pillow            11.2.1
pip               24.3.1
PyAutoGUI         0.9.54
PyGetWindow       0.0.9
PyMsgBox          1.0.9
pyperclip         1.9.0
PyRect            0.2.0
PyScreeze         1.0.1
python-docx       1.1.2
pytweening        1.2.0
typing_extensions 4.13.2
